Bento: Split amp-fit-text into new file structure
Per https://docs.google.com/document/d/1uJdH836Q8U5nhtCfVWc24kha2BNxU-Nkp-5riSgyI3k/edit, we'll need to split AMP components into separate files so they may be individually compiled into the npm package's modules.
