+++ extensions/amp-fit-text/1.0/amp-fit-text.js
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
+import {BaseElement} from './base-element';
 import {CSS} from '../../../build/amp-fit-text-1.0.css';
-import {FitText} from './fit-text';
-import {PreactBaseElement} from '../../../src/preact/base-element';
 import {isExperimentOn} from '../../../src/experiments';
 import {userAssert} from '../../../src/log';
 
 /** @const {string} */
 const TAG = 'amp-fit-text';
 
-class AmpFitText extends PreactBaseElement {
+class AmpFitText extends BaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -34,21 +33,6 @@ class AmpFitText extends PreactBaseElement {
     return super.isLayoutSupported(layout);
   }
 }
-
-/** @override */
-AmpFitText['Component'] = FitText;
-
-/** @override */
-AmpFitText['props'] = {
-  'minFontSize': {attr: 'min-font-size', type: 'number'},
-  'maxFontSize': {attr: 'max-font-size', type: 'number'},
-};
-
-/** @override */
-AmpFitText['passthrough'] = true;
-
-/** @override */
-AmpFitText['layoutSizeDefined'] = true;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpFitText, CSS);
