+++ extensions/amp-fit-text/1.0/base-element.js
@@ -1,0 +1,35 @@
+/**
+ * Copyright 2020 The AMP HTML Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS-IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import {FitText} from './component';
+import {PreactBaseElement} from '../../../src/preact/base-element';
+
+export class BaseElement extends PreactBaseElement {}
+
+/** @override */
+BaseElement['Component'] = FitText;
+
+/** @override */
+BaseElement['props'] = {
+  'minFontSize': {attr: 'min-font-size', type: 'number'},
+  'maxFontSize': {attr: 'max-font-size', type: 'number'},
+};
+
+/** @override */
+BaseElement['passthrough'] = true;
+
+/** @override */
+BaseElement['layoutSizeDefined'] = true;
