+++ extensions/amp-fit-text/1.0/component.js
@@ -15,7 +15,7 @@
  */
 
 import * as Preact from '../../../src/preact';
-import * as styles from './fit-text.css';
+import * as styles from './component';
 import {ContainWrapper} from '../../../src/preact/component';
 import {px, resetStyles, setStyle, setStyles} from '../../../src/style';
 import {toWin} from '../../../src/types';
