+++ extensions/amp-fit-text/1.0/component.type.js
